IDEA-179716 Cannot return Integer from method returning short. Fix tests.
(cherry picked from commit 1a7e33d)

+++ java/java-psi-impl/src/com/intellij/psi/impl/source/tree/java/PsiLiteralExpressionImpl.java
@@ -126,7 +126,7 @@ public class PsiLiteralExpressionImpl
       return PsiLiteralUtil.parseFloat(text);
     }
     if (type == JavaTokenType.DOUBLE_LITERAL) {
-      return PsiLiteralUtil.parseFloat(text);
+      return PsiLiteralUtil.parseDouble(text);
     }
     if (type == JavaTokenType.CHARACTER_LITERAL) {
       if (StringUtil.endsWithChar(text, '\'')) {
